--- src/main/java/step1/Calculator.java
@@ -1,19 +1,0 @@
-package step1;
-
-public class Calculator {
-    public int add(int left, int right) {
-        return 0;
-    }
-
-    public int subtract(int left, int right) {
-        return 0;
-    }
-
-    public int multiply(int left, int right) {
-        return 0;
-    }
-
-    public int divide(int left, int right) {
-        return 0;
-    }
-}
